requirements.txt
python-telegram-bot==21.5
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
